requirements.txt
pandas
numpy
pymysql
psutil
python-dotenv
requests
pymongo
